backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
numpy
scipy
sympy
pytest
python-dotenv
